+++ app/src/main/java/com/zybooks/foodrecommender/ui/RestaurantDetailViewModel.kt
@@ -5,5 +5,5 @@ import com.zybooks.foodrecommender.data.Restaurant
 import com.zybooks.foodrecommender.data.RestaurantDataSource
 
 class RestaurantDetailViewModel : ViewModel() {
-    fun getRestaurant(id: Int) : Restaurant = RestaurantDataSource().getRestaurant(id) ?: Restaurant()
+    fun getRestaurant(id: Long) : Restaurant = RestaurantDataSource().getRestaurant(id) ?: Restaurant()
 }
